+++ full-moon/src/ast/parsers.rs
@@ -1767,7 +1767,7 @@ fn parse_primary_expression(state: &mut ParserState) -> ParserResult<Expression>
                     ))
                 }
 
-                other => unreachable!("unexpected interpolated string kind: {other:?}"),
+                _ => ParserResult::NotFound,
             }
         }