+++ CHANGELOG.md
@@ -11,7 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Luau: support parsing function attributes (rfc: https://github.com/luau-lang/rfcs/blob/master/docs/syntax-attributes-functions.md)
 
 ### Changed
-- **[BREAKING CHANGE]** Introduced a new struct `AnonymousFunction` that is used in the `Expression::Function` variant. This replaces the old `Box<(TokenReference, FunctionBody)>`, and allows us to attach more data to an anonymous function (at the moment, mainly Luau attributes)
+- **[BREAKING CHANGE]** Introduced a new struct `AnonymousFunction` that is used in the `Expression::Function` variant. The old `Box<(TokenReference, FunctionBody)>` is now replaced with `Box<AnonymousFunction>`, and allows us to attach more data to an anonymous function (at the moment, mainly Luau attributes)
 
 ### Fixed
 - Fixed a panic when parsing a malformed empty interpolated string, e.g. ``print(`{}`)``